+++ sperl_allocator_runtime.h
@@ -3,6 +3,6 @@
 #include <stddef.h>
 #include "sperl_base.h"
 
-void* SPerl_ALLOCATOR_RUNTIME_alloc(SPerl* sperl, int64_t size);
+void* SPerl_ALLOCATOR_RUNTIME_alloc(SPerl* sperl, size_t size);
 
 #endif
